requirements.txt
blinker==1.9.0certifi==2025.8.3charset-normalizer==3.4.3click==8.2.1colorama==0.4.6Flask==3.0.3gunicorn==21.2.0idna==3.10itsdangerous==2.2.0Jinja2==3.1.6MarkupSafe==3.0.2packaging==25.0python-dotenv==1.0.1requests==2.32.3stripe==12.5.0typing_extensions==4.15.0urllib3==2.5.0Werkzeug==3.1.3psycopg2-binary==2.9.9Flask>=3.0,<4requests>=2.31,<3python-dotenv>=1.0,<2# Optional (nur nutzen, wenn Amazon-PAAPI aktiviert ist)amazon-paapi>=5,<6# Optional (nur wenn Stripe genutzt wird)stripe>=7,<8# Optional (nur wenn du Postgres statt SQLite verwendest)psycopg2-binary>=2.9,<3# Produktions-Server (unter Linux/macOS; auf Windows wird das ignoriert)gunicorn>=21,<22 ; sys_platform != "win32"